chore: rename earbydriversdk -> nearbydriver and NearbyDriverSDK -> NearbyDriver
