+++ js/android/app/src/main/java/tech/berty/gobridge/nearbydriver/util/BertyLogger.java
@@ -1,4 +1,4 @@
-package tech.berty.gobridge.nearbydriversdk.util;
+package tech.berty.gobridge.nearbydriver.util;
 
 
 import tech.berty.android.BuildConfig;
